feature(TableResponse): ported changes into the new DB-specific hierarchy
